Clean up DefaultAuthorService
